Add files via upload
Implemented a generic self-organizing binary search tree that can restructure itself depending on the search frequency of elements in the tree. Each time an item is searched, it receives a count, and when it passes a certain threshold, it will move up the tree by one level by a single rotation.