Add files via upload
Implemented a hash table class template as an adaptor class, then developed a simple spell checker. 
The program would input a file and detect if words are valid (in the hash table), then when an invalid word is detected, it will find the top 10 candidate words that could replace it.